requirements.txt
torch
torchvision
transformers
diffusers
accelerate
pillow
